requirements.txt
black
Django>=2
django-ipware>=2
django-user-agents<1
ipdb==0.13.4
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-factoryboy==2.1.0
pytest-pythonpath==0.7.3
